+++ src/YMM4VocaloidBridge.Core/Sequence/ISequencePlanner.cs
@@ -1,0 +1,8 @@
+using YMM4VocaloidBridge.Core.Reading;
+
+namespace YMM4VocaloidBridge.Core.Sequence;
+
+public interface ISequencePlanner
+{
+    SequencePlan Plan(JapaneseReadingResult reading, BridgeOptions options);
+}